requirements.txt
APScheduler==3.6.3
asgiref==3.6.0
attrs==23.1.0
beautifulsoup4==4.11.1
blinker==1.6.2
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.2.0
clarifai-grpc==7.11.0
click==8.1.7
Django==4.1.6
dlib==19.24.1
emoji==2.8.0
ephem @ file:///home/builder/ci_310/ephem_1640789806723/work
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.3.3
google-auth==2.12.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.60.0
grpcio==1.58.0
gspread==5.5.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
packaging==23.1
pluggy==1.3.0
protobuf==4.24.3
psycopg2-binary==2.9.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
pytest==7.0.1
pytest-sugar==0.9.4
python-telegram-bot==13.15
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
termcolor==2.3.0
tomli==2.0.1
tornado==6.1
tzlocal==5.0.1
urllib3==2.0.5
Werkzeug==2.3.7
